added reset after timer, added array to reset logic

+++ Assets/Scripts/LevelManager.cs
@@ -5,5 +5,5 @@ using System.Text;
 
 public static class LevelManager
 {
-    public static int level = 1;
+    public static float level = 1f;
 }
